A few post-review tweaks
